AI/Nova_Caster/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
llama-cpp-python>=0.2.0
customtkinter>=5.2.0